requirements.txt
openai>=1.50.0,<1.52.0
python-dotenv==1.0.1
httpx[http2]>=0.28.0